Solve using the piqp sparse matrix interface

Solve using the piqp sparse matrix interface

piqp_sparse_solve(P, c, A, b, G, h, x_lb, x_ub, rs)

Arguments

P

a symmetric positive definite matrix

c

the objective cost

A

the coefficient matrix of primal constraints

b

the right hand side vector of the primal constraints

G

the matrix of inequality constraints

h

the right side vector of the matrix inequality constraints

x_lb

the lower bounds for the solution

x_ub

the upper bounds for the solution

rs

a list of control settings

Value

named list

x (primal solution)

y (dual solution of equality constraints)

z (dual solution of inequality constraints)

z_lb (dual solution of lower bound box constraints)

z_ub (dual solution of upper bound box constraints)

obj (primal objective value)

run_time (total runtime, if so specified in settings)