Interface to 'piqp', a proximal interior point quadratic solver

PIQP solves linear programs (LPs) and quadratic programs (QPs). ,

Minimize $$\frac{1}{2}x^TPx + c^Tx$$ subject to $$Ax = b$$, $$Gx \leq h$$, $$x_{lb} \leq x \leq x_{ub}$$, where \(x \in R^n\), \(h \in R^m\), \(P = P^T\) and nonnegative-definite, \(c \in R^n\), \(A \in R^{p\times n}\), \(G \in R^{m\times n}\), \(b \in R^p\), \(x{lb} \in R^n\), \(x_{ub} \in R^n\).

piqp(
  A = matrix(0, nrow = 0, ncol = 0),
  b = numeric(0),
  c = numeric(0),
  P = matrix(0, nrow = 0, ncol = 0),
  G = matrix(0, nrow = 0, ncol = 0),
  h = numeric(0),
  x_lb = NULL,
  x_ub = NULL,
  control = list(),
  interface = c("auto", "dense", "sparse")
)

Arguments

A

a matrix of constraint coefficients

b

a numeric vector giving the primal constraints

c

a numeric vector giving the primal objective

P

a symmetric positive semidefinite matrix

G

a matrix of inequality coefficients

h

a numeric vector of the inequality right hand side

x_lb

a numeric vector of upper bounds, default NULL indicating Inf for all variables, otherwise should be number of variables long

control

a list giving specific control parameters to use in place of default values, with empty list indicating defaults

interface

whether to use sparse, or dense interface. Default is "auto" mode will switch to sparse interface if any of the specified matrices are sparse as in the Matrix package.

Value

named list

status (status code)

x (primal solution)

y (dual solution of equality constraints)

z (dual solution of inequality constraints)

z_lb (dual solution of lower bound box constraints)

z_ub (dual solution of upper bound box constraints)

pobj (primal objective value)

dobj (dual objective value)

dgap (absolute duality gap absolute)

dgap_rel (relative duality gap)

run_time (total runtime, if so specified in settings)

Details

PIQP uses compressed column sparse (CSC) format internally, which is the Matrix::dgCMatrix in the Matrix package. So even P should be in CSC format, not taking advantage of symmetry for instance. This package accepts either Matrix::dgcMatrix-class class objects or the triplet format sparse matrices as in the slam::simple_triplet_matrix()

See also